WareHouse/ImgEditor_THUNLP_20230824142740/requirements.txt
tkinter==8.6
Pillow==8.3.2
